rag_service/requirements.txt
# GigaChat API
gigachat==0.1.36

# Environment
python-dotenv==1.0.1

# HTTP
httpx==0.27.2

# Document parsing
pypdf==5.0.1
python-docx==1.1.2

# Testing
pytest==9.0.3